feat(p0147h): split 6 oversize files into per-subdomain partials
Mechanical refactor of three ServiceCollectionExtensions (Application 359 lines,
Infrastructure 188, Server 242) into per-subdomain helper-method partials
(AddSkillRuntime, AddPipelineHandlers, AddTriageServices, AddSecurityHandlers,
AddBootstrapHandlers, AddContextBuilders, AddPipelineRuntime; AddTicketProviders,
AddSourceProviders, AddChatClients, AddOutputStrategies, AddSecurityScanners,
AddDialogue, plus ConfigLoading; AddRedis, AddWebhooks, AddLongRunning, AddIntent,
AddPlatformAdapters, AddJobSpawner). Public AddAgentSmith() / Infrastructure /
Server entry points stay single — only the internal organisation changes.

Three data files (ContextKeys 171, CommandNames 202, PipelinePresets 289) become
`partial class` with per-subdomain partial files: ContextKeys.Pipeline / Plan /
Skill / Security / Api / Bootstrap; CommandNames.Pipeline / Security / Api plus
CommandProgressLabels.cs; PipelinePresets.{FixBug, FixNoTest, AddFeature,
InitProject, MadDiscussion, LegalAnalysis, SecurityScan, ApiSecurityScan,
SkillManager, Autonomous}. Constants stay on the original class — every
existing ContextKeys.X / CommandNames.Y / PipelinePresets.Z caller compiles
unchanged (684 / 395 / 80 caller references preserved).

PipelinePresets static dictionary initialization moved into a static
constructor — C# does not guarantee static-field init order across partial-
class files, so the inline literal triggered 52 test failures
(IsSinglePhase returned true for fix-no-test because FixNoTest was null at
All's initialization time). Static constructors run after all field
initializers regardless of compiler file order.

All 6 originals now under the 120-line hard limit (47/52/72/41/45/117 lines);
all new partials ≤80 lines target except three at the natural domain boundary
(CommandNames.Pipeline 81, CommandProgressLabels 87, PipelinePresets.cs 117 —
all under 120). Zero behaviour change: 1689 + 66 = 1755 tests pass; build
warning-clean (5 pre-existing SandboxToolHost obsolete warnings unchanged).
