+++ .agentsmith/phases/done/p0147h-di-and-data-file-splits.yaml
@@ -1,0 +1,124 @@
+# yaml-language-server: $schema=../../phase-spec.schema.json
+phase: p0147h
+goal: "Six files over the 120-line hard limit hold mechanical violations:
+  three ServiceCollectionExtensions (358, 242, 194 lines), three data
+  files (ContextKeys 174, CommandNames 206, PipelinePresets 291). DI
+  files need helper-method splits per subdomain; data files need
+  per-subdomain partial classes or split files. Low risk, high
+  readability gain, mechanical execution."
+
+requires: []
+
+scope:
+  in: |
+    (1) DI extensions split into per-subdomain helper methods within the
+        same static class. Public AddAgentSmith() / AddAgentSmithServer()
+        / AddAgentSmithInfrastructure() entry points stay; internal
+        helpers (AddPipelineHandlers, AddTriageServices, AddSkillRuntime,
+        AddTicketProviders, AddSourceProviders, AddOutputStrategies,
+        AddSandbox, AddBus) each ≤30 lines.
+
+    (2) ContextKeys split by subdomain:
+        - PipelineContextKeys (RunId, RunStartedAt, CheckoutBranch, ...)
+        - SecurityContextKeys (SkillObservations, SecurityTrend, ...)
+        - ApiContextKeys (SwaggerSpec, ApiTarget, ...)
+        - BootstrapContextKeys (project_language, context_yaml_present, ...)
+        Either partial class or separate files exposing the same
+        constants on the existing ContextKeys class for caller-side
+        zero-impact.
+
+    (3) CommandNames split similarly: per-subdomain partial classes or
+        separate static classes (PipelineCommandNames, SecurityCommandNames,
+        ApiCommandNames). Progress-label map (today inline at line 159+)
+        moves to its own file CommandProgressLabels.cs (~60 lines).
+
+    (4) PipelinePresets: today static arrays for FixBug / FixNoTest /
+        AddFeature / SecurityScan / ApiSecurityScan / MadDiscussion /
+        LegalAnalysis / InitProject / Autonomous / SkillManager. Two
+        options to consider during implementation:
+        (a) One file per preset (PipelinePresets.FixBug.cs etc. as
+            partial classes), or
+        (b) Load presets from a YAML file at startup (deferred — bigger
+            change, separate phase).
+        Phase commits to (a) here; (b) tracked as a follow-up if it's
+        worth the operator-facing config-surface change.
+
+  out: |
+    Changing the DI registration semantics. Same services in the same
+    lifetimes; only the registration code shape changes.
+
+    Renaming any existing public constants. Caller-side zero-impact is
+    a hard requirement of this phase.
+
+decisions:
+  - key: |
+      Helper-method splits inside the existing static class, not
+      separate extension classes. Reason: AddAgentSmith() stays the
+      single public API surface; reviewers + operators see one entry
+      point. The internal organisation is for readability, not for
+      external composition.
+  - key: |
+      Constants stay on their original class via partial class. Reason:
+      every caller in the codebase references ContextKeys.X /
+      CommandNames.Y — a rename would touch hundreds of call sites and
+      add zero value. Partial class preserves the surface.
+  - key: |
+      Defer YAML-driven presets to a separate phase. Reason: that's a
+      semantic change (operator configurability) not a refactor; this
+      phase commits to mechanical splits only. Track as p0148-candidate
+      if it surfaces operator demand.
+
+steps:
+  - id: di-extensions-split
+    action: "Three DI extension classes get helper-method splits. AddAgentSmith() / equivalent stays as the entry point and calls each helper."
+    modify:
+      - "src/AgentSmith.Application/Extensions/ServiceCollectionExtensions.cs"
+      - "src/AgentSmith.Infrastructure/Extensions/ServiceCollectionExtensions.cs"
+      - "src/AgentSmith.Server/Extensions/ServiceCollectionExtensions.cs"
+
+  - id: context-keys-split
+    action: "ContextKeys becomes partial; per-subdomain partial files."
+    modify:
+      - "src/AgentSmith.Contracts/Commands/ContextKeys.cs (now partial, top-level shared keys only)"
+    new:
+      - "src/AgentSmith.Contracts/Commands/ContextKeys.Pipeline.cs"
+      - "src/AgentSmith.Contracts/Commands/ContextKeys.Security.cs"
+      - "src/AgentSmith.Contracts/Commands/ContextKeys.Api.cs"
+      - "src/AgentSmith.Contracts/Commands/ContextKeys.Bootstrap.cs"
+
+  - id: command-names-split
+    action: "CommandNames becomes partial; progress-label map extracted to its own class."
+    modify:
+      - "src/AgentSmith.Contracts/Commands/CommandNames.cs (partial)"
+    new:
+      - "src/AgentSmith.Contracts/Commands/CommandProgressLabels.cs"
+      - "src/AgentSmith.Contracts/Commands/CommandNames.Api.cs"
+      - "src/AgentSmith.Contracts/Commands/CommandNames.Security.cs"
+      - "src/AgentSmith.Contracts/Commands/CommandNames.Pipeline.cs"
+
+  - id: pipeline-presets-split
+    action: "PipelinePresets becomes partial; per-preset partial files."
+    modify:
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.cs (partial — TryResolve + presetNames + GetPipelineType)"
+    new:
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.FixBug.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.SecurityScan.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.ApiSecurityScan.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.MadDiscussion.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.AddFeature.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.InitProject.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.LegalAnalysis.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.Autonomous.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.SkillManager.cs"
+      - "src/AgentSmith.Contracts/Commands/PipelinePresets.FixNoTest.cs"
+
+tests:
+  - "Existing test suite passes (zero behavioural change)."
+  - "Each split file ≤80 lines."
+  - "Caller-side: a grep for ContextKeys.X / CommandNames.Y / PipelinePresets.Z returns identical match sets pre and post-refactor."
+
+done:
+  - "All six target files (3 DI + 3 data) are under the 120-line hard limit."
+  - "Each partial / helper file ≤80 lines."
+  - "Caller-side zero-impact verified."
+  - "Full test suite green."
