+++ src/AgentSmith.Server/Extensions/ServiceCollectionExtensions.LongRunning.cs
@@ -1,0 +1,36 @@
+using AgentSmith.Contracts.Services;
+using AgentSmith.Server.Services;
+using AgentSmith.Server.Services.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace AgentSmith.Server.Extensions;
+
+internal static partial class ServiceCollectionExtensions
+{
+    // Long-running hosted services: queue consumer, housekeeping leader, poller leader,
+    // and the Redis connection health monitor. Each one is published as ISubsystemHealth
+    // so /health endpoints expose the per-subsystem status.
+    internal static IServiceCollection AddLongRunningServices(this IServiceCollection services)
+    {
+        services.AddSingleton<QueueConsumerHostedService>();
+        services.AddHostedService(sp => sp.GetRequiredService<QueueConsumerHostedService>());
+        services.AddSingleton<ISubsystemHealth>(sp =>
+            sp.GetRequiredService<QueueConsumerHostedService>().Health);
+
+        services.AddSingleton<HousekeepingLeaderHostedService>();
+        services.AddHostedService(sp => sp.GetRequiredService<HousekeepingLeaderHostedService>());
+        services.AddSingleton<ISubsystemHealth>(sp =>
+            sp.GetRequiredService<HousekeepingLeaderHostedService>().Health);
+
+        services.AddSingleton<PollerLeaderHostedService>();
+        services.AddHostedService(sp => sp.GetRequiredService<PollerLeaderHostedService>());
+        services.AddSingleton<ISubsystemHealth>(sp =>
+            sp.GetRequiredService<PollerLeaderHostedService>().Health);
+
+        services.AddSingleton<RedisConnectionHealth>();
+        services.AddSingleton<ISubsystemHealth>(sp =>
+            sp.GetRequiredService<RedisConnectionHealth>().Health);
+
+        return services;
+    }
+}
