+++ src/AgentSmith.Server/Extensions/ServiceCollectionExtensions.Webhooks.cs
@@ -1,0 +1,31 @@
+using AgentSmith.Contracts.Services;
+using AgentSmith.Server.Services.Handlers;
+using AgentSmith.Server.Services.Webhooks;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace AgentSmith.Server.Extensions;
+
+internal static partial class ServiceCollectionExtensions
+{
+    // p0140b: shared per-match spawn loop + zero-match handler used by all 8
+    // ticket-event handlers. Each platform (GitHub/GitLab/AzDo/Jira) contributes its
+    // own IWebhookHandler(s) for the issue / comment / PR-label / PR-comment event flavours.
+    internal static IServiceCollection AddWebhookHandlers(this IServiceCollection services)
+    {
+        services.AddSingleton<WebhookSpawnDispatcher>();
+        services.AddSingleton<IWebhookHandler, GitHubIssueWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitHubIssueCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitHubPrLabelWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitHubPrCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitLabIssueWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitLabIssueCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitLabMrLabelWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, GitLabMrCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, AzureDevOpsWorkItemWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, AzureDevOpsWorkItemCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, AzureDevOpsPrCommentWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, JiraAssigneeWebhookHandler>();
+        services.AddSingleton<IWebhookHandler, JiraCommentWebhookHandler>();
+        return services;
+    }
+}
